Move main class to its own class file
